More tidy up work on win32 converter
