+++ tool/win32/convert.dart
@@ -3,4 +3,5 @@ import 'shared.dart';
 void main() {
   loadCsv('tool/win32/win32api.csv');
   saveCsv('tool/win32/win32api.csv');
+  saveJson('tool/win32/win32api.json');
 }
